requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
beautifulsoup4==4.12.3
undetected-chromedriver==3.5.5
tldextract==5.1.2
python-dotenv==1.0.1
setuptools==69.2.0
apscheduler==3.10.4
